исправления, дополнения, приведение к общему формату

+++ src/main/java/com/vstavaystrana/vstavaystrana_site/models/AbstractEntity.java
@@ -9,10 +9,10 @@ public class AbstractEntity implements DatabaseEntity
 {
         @Id
         @GeneratedValue
-        protected Integer id;
+        protected Long id;
 
         @Override
-        public Integer getId()
+        public Long getId()
         {
             return id;
         }
