+++ src/main/java/com/vstavaystrana/vstavaystrana_site/models/Buisnessman.java
@@ -5,7 +5,7 @@ import jakarta.persistence.Entity;
 @Entity
 public class Buisnessman extends AbstractEntity
 {
-    protected Integer userId;
+    protected Long userId;
     protected Byte[] passportScan;
     protected Float readiness;
     protected Boolean haveCompletedProject;
