+++ src/main/java/com/vstavaystrana/vstavaystrana_site/models/News.java
@@ -5,7 +5,7 @@ import jakarta.persistence.Entity;
 @Entity
 public class News extends AbstractEntity
 {
-    protected Integer projectId;
+    protected Long projectId;
     protected String author;
     protected String content;
     protected String photoUrl;
